Add vector icon as peer dependency and show chevron-right icon on the picker box
