+++ example/package.json
@@ -17,6 +17,7 @@
     "react-native": "0.70.5",
     "react-native-gesture-handler": "^2.6.0",
     "react-native-reanimated": "~2.12.0",
+    "react-native-vector-icons": "^9.2.0",
     "react-native-web": "~0.18.9"
   },
   "devDependencies": {
